# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/chat/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "CreateChat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/get-search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "GetSearch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chatID",
                        "name": "chatID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "messageID",
                        "name": "messageID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.ChatActivity"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/search-bkw": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "SearchByKeyWord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chatID",
                        "name": "chatID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.ChatActivity"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/x-to-y": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "UpdateAvatarAsync",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "oldAvatar",
                        "name": "oldAvatar",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "newAvatar",
                        "name": "newAvatar",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/group/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "CreateGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/group/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "GetGroupInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "groupID",
                        "name": "groupID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Group"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentications"
                ],
                "summary": "CreateUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/info/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateGroupReq": {
            "type": "object",
            "properties": {
                "admins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PersonInfo"
                    }
                },
                "chatAvatar": {
                    "type": "string"
                },
                "chatName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PersonInfo"
                    }
                },
                "owner": {
                    "$ref": "#/definitions/entity.PersonInfo"
                },
                "setting": {
                    "$ref": "#/definitions/entity.GroupSetting"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.HTTPResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "entity.ChatActivity": {
            "type": "object",
            "properties": {
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Content"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "hidden": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "messageID": {
                    "type": "string"
                },
                "parentID": {
                    "type": "string"
                },
                "recall": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "entity.Content": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "entity.Conversation": {
            "type": "object",
            "properties": {
                "chatAvatar": {
                    "type": "string"
                },
                "chatId": {
                    "type": "string"
                },
                "chatName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deliveries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Delivery"
                    }
                },
                "id": {
                    "type": "string"
                },
                "id_UserOrGroup": {
                    "type": "string"
                },
                "reads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Delivery"
                    }
                },
                "topChatActivities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ChatActivity"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.Delivery": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messageID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "entity.FriendRequest": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isSender": {
                    "type": "boolean"
                },
                "sendAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "entity.Group": {
            "type": "object",
            "properties": {
                "admins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PersonInfo"
                    }
                },
                "chatAvatar": {
                    "type": "string"
                },
                "chatName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PersonInfo"
                    }
                },
                "owner": {
                    "$ref": "#/definitions/entity.PersonInfo"
                },
                "setting": {
                    "$ref": "#/definitions/entity.GroupSetting"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.GroupSetting": {
            "type": "object",
            "properties": {
                "changeChatNameAndAvatar": {
                    "type": "boolean"
                },
                "createNewPolls": {
                    "type": "boolean"
                },
                "membershipApproval": {
                    "type": "boolean"
                },
                "pinMessages": {
                    "type": "boolean"
                },
                "sendMessages": {
                    "type": "boolean"
                }
            }
        },
        "entity.PersonInfo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userAvatar": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "entity.User": {
            "type": "object",
            "properties": {
                "conversations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Conversation"
                    }
                },
                "friendRequests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FriendRequest"
                    }
                },
                "id": {
                    "type": "string"
                }
            }
        }
    }
}